CANTools.Utils

CANTools.Utils.from_bytes — Method

from_bytes(type::Type{T}, array::AbstractArray{UInt8}) where {T <: Number} -> T

Creates a value of type T constituted by the byte-array array. If the array length is smaller than the size of T, array is filled with enough zeros.

Arguments

type::Type{T}: the type to which the byte-array is transformed

array::AbstractArray{UInt8}: the byte array

Returns

T: the value constructed from the byte sequence

Examples

using CANTools.Utils
bytes = Utils.from_bytes(UInt16, UInt8[0xFF, 0xAA])

# output
0xaaff

CANTools.Utils.full_mask — Method

full_mask(::Type{T}) where {T <: Integer} -> T

Creates a full mask of type T with 8sizeof(T) bits.

Arguments

Type{T}: the type of the mask

Returns

T: the full mask

Examples

using CANTools.Utils
m = Utils.full_mask(Int8)

# output
-1

CANTools.Utils.is_big_endian — Method

is_big_endian() -> Bool

Returns whether the system has big-endian byte-order

Returns

Bool: The system has big-endian byte-order

CANTools.Utils.is_little_endian — Method

is_little_endian() -> Bool

Returns whether the system has little-endian byte-order

Returns

Bool: The system has little-endian byte-order

CANTools.Utils.mask — Method

mask(::Type{T}) where {T <: Integer} -> T

Creates a full mask of type T with 8sizeof(T) bits.

Arguments

Type{T}: the type of the mask

Returns

T: the full mask

Examples

using CANTools.Utils
m = Utils.mask(UInt64)

# output
0xffffffffffffffff

CANTools.Utils.mask — Method

mask(::Type{T}, length::Integer, shift::Integer) where {T <: Integer} -> T

Creates a mask of type T with length number of bits and right-shifted by shift number of bits.

Arguments

Type{T}: the type of the mask

length::Integer: the number of bits

shift::Integer: the right-shift

Returns

T: the mask defined by length and shift

Examples

using CANTools.Utils
m = Utils.mask(UInt64, 32, 16)

# output
0x0000ffffffff0000

CANTools.Utils.mask — Method

mask(::Type{T}, length::Integer) where {T <: Integer} -> T

Creates a mask of type T with length number of bits.

Arguments

Type{T}: the type of the mask

length::Integer: the number of bits

Returns

T: the mask defined by length

Examples

using CANTools.Utils
m = Utils.mask(UInt64, 32)

# output
0x00000000ffffffff

CANTools.Utils.mask — Method

mask(::Type{T}, length::UInt8, shift::UInt8) where {T <: Integer} -> T

Creates a mask of type T with length number of bits and right-shifted by shift number of bits.

Arguments

Type{T}: the type of the mask

length::UInt8: the number of bits

shift::UInt8: the right-shift

Returns

T: the mask defined by length and shift

CANTools.Utils.mask — Method

mask(::Type{T}, length::UInt8) where {T <: Integer} -> T

Creates a mask of type T with length number of bits.

Arguments

Type{T}: the type of the mask

length::UInt8: the number of bits

Returns

T: the mask defined by length

CANTools.Utils.to_bytes — Method

to_bytes(num::Number) -> Vector{UInt8}

Creates the byte representation of the number num.

Arguments

num::Number: the type of the mask

Returns

Vector{UInt8}: the bytes representation of the number num

Examples

using CANTools.Utils
bytes = Utils.to_bytes(UInt16(0xAAFF))

# output
2-element Vector{UInt8}:
 0xff
 0xaa

CANTools.Utils.zero_mask — Method

zero_mask(::Type{T}) where {T <: Integer} -> T

Creates a zero mask of type T where every bit is unset.

Arguments

Type{T}: the type of the mask

Returns

T: the zero mask

Examples

using CANTools.Utils
m = Utils.zero_mask(UInt8)

# output
0x00